manshu-python/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
aiomysql==0.2.0
PyMySQL==1.1.0

# Redis
redis[hiredis]==5.0.1

# MinIO 对象存储
minio==7.2.0

# Elasticsearch
elasticsearch==8.10.0

# Kafka
aiokafka==0.10.0

# OpenAI & LangChain
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 验证码（只需要 Pillow）
Pillow>=10.0.0

# 邮件
aiosmtplib==3.0.1

# HTTP 客户端
httpx==0.25.2

# 工具
python-dateutil==2.8.2

